refactor(pkg/server): move database to new package
