+++ pkg/database/database_protoconv.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package database
 
 import (
 	"crypto/sha256"
