+++ pkg/database/dboptions.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package database
 
 //DbOptions database instance options
 type DbOptions struct {
@@ -30,7 +30,7 @@ func DefaultOption() *DbOptions {
 	return &DbOptions{
 		//		dbDir:             "immudb",
 		dbName:            "db_name",
-		dbRootPath:        DefaultOptions().Dir,
+		dbRootPath:        "./data",
 		corruptionChecker: true,
 		inMemoryStore:     false,
 	}
